+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 * Initial Release
 
 # 0.0.2
-* Added some more data to the output, such as instanceid
+* Added some more data to the output, such as instance_id
 
 # 0.0.3
 
 * Output is now JSON rather then a ruby hash
+
+# 0.0.4 
+
+* Supports returning only a single value, such as "host", "instance_id", ect